requirements.txt
streamlit
python-docx
openai
python-dotenv
pillow
firebase-admin
pandas
SpeechRecognition 
gTTS 
audio-recorder-streamlit 
pydub
